refactor(security): authenticationManager

+++ src/main/java/com/tid/avisExpress/repository/ValidationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ValidationRepository extends CrudRepository<Validation, Integer> {
+public interface ValidationRepository extends CrudRepository<Validation, Long> {
     Optional<Validation> findByCode(String code);
 }
